build(kafka-vendoring): relocate and clarify Kafka vendoring tooling
Rename kroxylicious-api/etc/kafka-support to top-level tools/kafka-vendoring:
it's a standalone script-based tool whose only tie to kroxylicious-api is
where its output lands, closer in kind to the existing top-level scripts/
than to a module's own etc/ config.

Rename refresh.sh/surgery.py/closure.txt to vendor.sh/apply-edits.py/
vendored-files.txt so the names say what they do. Replace surgery.py's
per-file if/elif dispatch with a generic apply-edits.py engine driven by a
new declarative edits.yaml, removing the SURGERY bash variable that had to
be hand-kept in sync with the Python dispatch. Extract the pom heredoc'd
inside refresh.sh into a tracked, diffable rewrite-pom.xml.

vendor.sh now also runs `mvn -pl kroxylicious-api -am process-sources` as
its final step, so its output is the same formatter/impsort-normalised form
that gets committed, rather than leaving that as a manual follow-up.

Re-ran the tool end-to-end against the pinned Kafka 4.3.0 checkout: zero
diff against the currently committed vendored tree.
