+++ tools/kafka-vendoring/requirements.txt
@@ -1,0 +1,5 @@
+# Copyright Kroxylicious Authors.
+#
+# Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+#
+PyYAML==6.0.2
