+++ tools/kafka-vendoring/rewrite-pom.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright Kroxylicious Authors.
+
+    Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+
+-->
+<!--
+    Throwaway pom used by vendor.sh to invoke the OpenRewrite ChangePackage recipe (rewrite.yml)
+    against a temporary staging copy of Kafka source. This is NOT a reactor module: it has no
+    source of its own and must never build as part of `mvn verify` at the repo root. vendor.sh
+    copies this file into its mktemp staging directory (alongside the freshly copied and edited
+    Kafka sources) and invokes rewrite-maven-plugin from there.
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>io.kroxylicious.tools</groupId>
+    <artifactId>kafka-vendoring-rewrite</artifactId>
+    <version>1.0</version>
+    <packaging>jar</packaging>
+
+    <properties>
+        <maven.compiler.release>21</maven.compiler.release>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <!-- rewrite:run forks the lifecycle through 'compile'. The staged sources deliberately do
+             not compile here (they still reference org.apache.kafka before relocation, and native
+             codec libs are off this throwaway classpath). ChangePackage is textual and needs no
+             type attribution, so we let the compile attempt fail without aborting the build. -->
+        <maven.compiler.failOnError>false</maven.compiler.failOnError>
+    </properties>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.openrewrite.maven</groupId>
+                <artifactId>rewrite-maven-plugin</artifactId>
+                <version>6.46.1</version>
+                <configuration>
+                    <activeRecipes>
+                        <recipe>io.kroxylicious.RewriteCopiedKafka</recipe>
+                    </activeRecipes>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+</project>
