+++ tools/kafka-vendoring/vendored-files.txt
@@ -1,3 +1,11 @@
+# Copyright Kroxylicious Authors.
+#
+# Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+#
+# The list of files copied from <kafka-checkout>/clients/src/main/java/org/apache/kafka/common
+# by vendor.sh. Computed once as a compiler-driven closure (javac with kafka-clients off the
+# classpath) so the vendored tree is self-contained under io.kroxylicious.kafka.common.* with no
+# org.apache.kafka.* references left. See README.md for how it was derived and when to recompute it.
 InvalidRecordException.java
 KafkaException.java
 Uuid.java
